separating tile.c in sub directory priv/tile/
